- Jenis sampah is now fetch from database - TODO: Fix the post on setor sampah

+++ lib/features/trash_bank/models/deposit_asus_model.dart
@@ -12,7 +12,7 @@ class DepositAsusModel {
   });
 
   final String desaId, berat, jenisSampah, rt, rw, userId;
-  final int poin;
+  final double poin;
   final DateTime waktu;
   final bool available;
